backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-socketio==5.10.0
python-dotenv==1.0.0
easyocr==1.7.0
pytesseract==0.3.10
opencv-python==4.8.1.78
mss==9.0.1
pillow==10.1.0
numpy==1.24.3
aiohttp==3.9.1
beautifulsoup4==4.12.2
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
ebaysdk==2.2.0